Added lookup based explict PLL Patch
Expanded on the existing explicit PLL patch by incorporating a lookup table
to allow an arbitrary number of specified fixed PLL configurations to be
inspected as needed, and not needing to add if blocks for each. This
patch is intended to be INSTEAD OF the existing AD9081 12Ghz/333Mhz
patch, which will be removed in a future commit.